Split the tarcache tests up into black and whitebox
